Create example_visualize_networks.py
example_visualize_networks.py is a Python script that showcases a basic implementation of visualizing a network using visdcc and Dash.

This script showcases things like how to create and deploy a Dash app, how to define node and edges of a network, and how to create/display a visdcc Network on the Dash app.