Added new prediction cells to twin_network_for_diffraction.ipynb
The new prediction cells would make predictions with the embedding model by comparing an unknown image to an anchor single-hit image and an anchor multi-hit image. If the distance between the embeddings of the unknown image and anchor single-hit image is smaller than the distance between the embeddings of the unknown image and anchor multi-hit image, the unknown image is a single-hit image. Otherwise the unknown image is a multi-hit image.

predict_using_anchors() takes in an anchor image for single-hit, an anchor image for multi-hit, and an unknown image. Returns a prediction for the unknown image based on the distances between the multi-hit and single-hit vectors.

sample_anchor_and_unknown_images() takes in list of images and their labels, and returns a sampled anchor image for single-hit image, anchor image for multi-hit image, and unknown image.