Fixed small bug in setup.py and utils.py.
Fixed a tiny bug in setup.py, which was some missing commas in the "requirements" list in the script.

Fixed another tiny bug in utils.py in the check_img_for_nan() function. The previous way of searching for 'nan' values, which was "if np.nan in img", was said to be incorrect online. The new version does the job properly.